Add a Windows console display
The display uses the Windows Console API [1] to output Unicode chess
symbols and color the board appropriately.

We could use ANSI escape sequences, which would work on recent versions
of Windows and typical POSIX systems. But, consistent Unicode output is
problematic, so for the time being, the display only supports Windows.

[1]: https://learn.microsoft.com/en-us/windows/console/console-reference