Be more "generating-oriented"
Let me explain. I imagine a "checking-oriented" chess implementation
is one where a move is described in full detail, and then the code
checks whether the move is allowed.

By generating-oriented, I mean less information about the move is
specified up front. We leave the code to generate all of the legal
moves that still satisfy our description.

So now, instead of specifying what we want a pawn to promote to, we
just say we want to move the pawn and the pawn-specific code returns
a list containing entries for every piece it could promote to.

I think this becomes more necessary as you consider chess variants.
In Chess960, for instance, it's possible for a king to castle by
moving only one square instead of two. So now, specifying where the
king is moving from and to isn't enough because the player may or may
not want to castle. Now we can just return a list containing both
possibilities and let the user choose one.