Be more "generating-oriented"
A "checking-oriented" chess implementation has code to check if a fully
specified move is legal or not. A "generating-oriented" implementation
generates several legal moves based on a less-specific description.

For example, instead of specifying the piece that a pawn should promote
to, we only specify where the pawn should move, and the code generates a
list containing one entry for each possible promotion.

This is an important step in supporting chess variants. In Chess960, for
instance, a king can sometimes castle by moving only one square instead
of two. Therefore, specifying the location of the king is not enough as
the player may or may not want to castle. In the future, we can simply
generate both moves, allowing the player to choose one.