Fix a castling bug
Consider a game in Chess960 where there is a white rook at 'c1' and a
white king at 'e1'. All of the conditions for castling are met.

What happens?

Kings always move to the 'c' file when long castling, so we move the
king from 'e1' to 'c1'. Oops! The rook was overwritten. When we
attempt to move the "rook" from 'c1' to 'd1', we actually just move
the king to 'd1'. Not only is the rook gone, but the king is in the
wrong spot!

The solution is pretty simple. Save or pick up the rook before moving
anything else. Move the king. Then restore the rook, i.e., put it back
down.

Another idea is to have a function that makes several moves
simultaneously. It basically picks all of the pieces up and then puts
them all back down, instead of moving one, then moving the next, then
moving another, and so on.