Fix a castling bug
Consider a Chess960 position where there is a white rook on 'c1' and a
white king on 'e1'. Assume that the conditions for castling are met.

Kings move to the 'c' file when long castling, so we move the king from
'e1' to 'c1', overwriting the rook! When we attempt to move the "rook"
from 'c1' to 'd1', we end up moving the king instead, leaving the king
on the wrong square and causing the rook to disappear.

There is a pretty simple fix: pick up the rook before moving anything;
move the king; and finally, put the rook down on the correct square.