Implement Chess960 :^)
I always wanted to support some variants, and Chess960 is a pretty
simple one. All of the rules are the same. The twist is that the
initial board is randomized. I implemented the so-called direct
derivation from Wikipedia to generate the board. See

https://en.wikipedia.org/wiki/Fischer_random_chess_numbering_scheme

for more details. One catch is that your castling code needs to be
more generic. Really, the rules are the same in classical chess and
Chess960. But in classical chess, you can safely assume, for instance,
that the black king will be on 'e8' before castling. That's not
necessarily the case in Chess960.