Fix an error from GCC 13.1 on Linux
According to GCC's documentation (see `-Wno-changes-meaning`), C++ does
not permit certain ambiguous names. More precisely, unqualified names in
a class must have the same meaning in the complete scope of the class.

For example, the declaration `color color;` in `Piece` is not allowed.
By defining the member variable `color`, the meaning of `color` changes
from an enumeration to a member variable. A direct fix is to prepend the
`enum` keyword, as in `enum color color;`, to clarify to the compiler
that we are referring to the enumeration named `color`.

The documentation goes on to say that:

> Both diagnostics are also suppressed by `-fms-extensions`.

I use Windows, and GCC on Windows seems to enable `-fms-extensions` by
default, so I never encountered that error until now!