Fix an error from GCC 13.1 on Linux
According to GCC's documentation (see `-Wno-changes-meaning`), C++ does
not allow certain ambiguous names. More precisely, unqualified names in
a class must have the same meaning in the complete scope of the class.

For example, the declaration `color color;` in the `Piece` structure
is not allowed. By defining the member variable `color`, we've changed
the meaning of that name from an enumeration to a member variable. The
simplest fix is to add the `enum` keyword before the declaration, as in
`enum color color;`. This clarifies that we are explicitly looking for
the enumeration named `color` and nothing else.

The documentation goes on to say that:

> Both diagnostics are also suppressed by `-fms-extensions`.

I've always compiled the project on Windows. On Windows, GCC is
apparently configured such that `-fms-extensions` is passed in by
default. So I never encountered the error until now!