Move TUI-specific logic into `TerminalUserInterface`
The `UserInterface` class now has user input methods as well as display
methods. For terminal UIs, I'm reusing the logic that was previously in
the `Game` class. These changes are in preparation for adding a GUI.

User interfaces must provide a way to choose a move out of a list of
legal moves (in case multiple moves have the same from/to square).

Additionally, user interfaces must provide a way to present notification
messages to the player. This is used to announce checkmate, stalemate,
check, and illegal moves. However, the `notify_pause` enum feels hacky.