Move more TUI-specific logic
The `UserInterface` class is now responsible for getting user input in
whichever way makes sense. For terminal UIs, I'm just reusing the logic
that was previously in `Game`. These changes are in preparation for
adding a GUI.

User interfaces must also provide a way to choose a move out of a list
of legal moves (in case multiple moves have the same from/to square).

Finally, user interfaces must provide a way to notify a user given some
notification message. This is only used to announce checkmate, stalemate,
check, and illegal moves. I feel like there's a better way...